Switch to yet another workaround for memchr on UEFI target
Bringing in Criterion (next commit) meant that it was no longer possible to pin memchr to version 2.1. Instead, use a cfg flag that forces memchr to avoid *all* SIMD optimizations. This could have a significant performance impact, but it probably won't matter much for the small amount of time we're in the UEFI environment.