Fix issue where LoadedImage was never fetched
This was thanks to some undefined behavior: I casted a reference to an immutable object into a mutable pointer. The loaded_image pointer was never updated, so it was always null. On x86_64, the system was happy to dereference a null pointer, so I didn't notice until I ported to ARM.